services/applications/adhoc-inference-server/requirements.txt
grpcio==1.62.1
protobuf==3.20.3
requests==2.32.3
redis==3.2.1
psycopg2-binary
flask==3.0.3